chore: move and rename file auth
