+++ internal/middleware/auth.go
@@ -1,4 +1,4 @@
-package auth
+package middleware
 
 import (
 	"fitbyte/internal/models"
